# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MailUp - Backend Test",
        "version": "0.0.1"
    },
    "paths": {
        "/api/v1/brands": {
            "get": {
                "tags": [
                    "Marcas"
                ],
                "summary": "Mostrar marcas existentes",
                "description": "Permite detallar los marcas registrados.",
                "operationId": "265f173d55c77d2ffffe46be4ba57fa8",
                "responses": {
                    "200": {
                        "description": "Operacion Exitosa."
                    }
                }
            },
            "post": {
                "tags": [
                    "Marcas"
                ],
                "summary": "Registrar nueva Marca",
                "description": "Permite registrar una nueva Marca.",
                "operationId": "354e334c26df48b4a20817df45f8fa7b",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Nombre de la Marca",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Sin autenticar."
                    },
                    "422": {
                        "description": "El campo ya esta en uso."
                    },
                    "403": {
                        "description": "Prohibido. No posee permisos para ejecutar esta accion."
                    }
                }
            }
        },
        "/api/v1/brands/{id}": {
            "get": {
                "tags": [
                    "Marcas"
                ],
                "summary": "Obtener información de una Marca",
                "description": "Devuelve datos de una Marca específica según Id. ",
                "operationId": "41ee41610cd373b7f6b3ecf7b8253877",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ingrese Id de una Marca existente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa."
                    },
                    "401": {
                        "description": "Sin autenticar."
                    },
                    "403": {
                        "description": "Prohibido. No posee permisos para ejecutar esta accion."
                    },
                    "404": {
                        "description": "Recurso no encontrado."
                    }
                }
            },
            "put": {
                "tags": [
                    "Marcas"
                ],
                "summary": "Actualizar una Marca existente",
                "description": "Permite modificar datos de una Marca existente.",
                "operationId": "818bae2e9412f13a474dd3af9c4049b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de Marca",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Nombre de la Marca",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Sin autenticar."
                    },
                    "422": {
                        "description": "El campo ya esta en uso."
                    },
                    "403": {
                        "description": "Prohibido. No posee permisos para ejecutar esta accion."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Marcas"
                ],
                "summary": "Eliminar Marca existente",
                "description": "Permite eliminar un registro y no devuelve contenido.",
                "operationId": "a5475e12bf803edae8884e43dca17735",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de Marca a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa."
                    },
                    "401": {
                        "description": "Sin autenticar."
                    },
                    "403": {
                        "description": "Prohibido. No posee permisos para ejecutar esta accion."
                    },
                    "404": {
                        "description": "Recurso no encontrado."
                    }
                }
            }
        },
...
    }
}